fix: ripemd160 library was broken on linux

+++ include/ripemd.c
@@ -1,4 +1,12 @@
-typedef unsigned long uint32;  // "unsigned int" could appear 16-bit on some older compilers
+// I AM NOT THE AUTHOR OF THIS FILE
+//
+// author: agoebel
+// github: https://github.com/agoebel/RIPEMD-160
+
+// author's note: "unsigned int" could appear 16-bit on some older compilers
+// my note: "unsigned long" actually appears as 64-bit for my tests on linux
+//          so I'm going with unsigned int
+typedef unsigned int uint32;
 
 #define BLOCK_LEN 64 // length of each block for processing in bytes
 #define HASH_LEN 20 // length of hash in bytes
